fix: build issue by renaming the file to onetrust
