+++ libs/plugins/rudder-plugin-onetrust-consent-filter-react-native/src/index.ts
@@ -1,4 +1,4 @@
-import OnetrustConsentFilter, { startConsentFilterPlugin } from './onetrustConsentFilterPlugin';
+import OnetrustConsentFilter, { startConsentFilterPlugin } from './onetrustFilterPlugin';
 
 export default OnetrustConsentFilter;
 export { startConsentFilterPlugin };
